fix: VS Code not inserting final newline
Previous to this change, VS Code was not inserting the final newline at
the end of the file.

There are two problems with this.

First, POSIX defines a line as a sequence of zero or more non-newline
characters plus a terminating newline character.  So this behavior was
not POSIX compliant.  See https://stackoverflow.com/a/729795.

More importantly, this was causing unnecessary diffs when editing files
on Linux.  Vim and others insert the final newline when saving a file
after making an unrelated change.

Fix by updating VS Code settings to insert the final newline.  See
https://stackoverflow.com/a/44704969.