feat: enable fetching from local webserver
using the Tauri fetch API

We can't fetch from GitHub Enterprise because it requires authentication
which we don't support.  As a workaround, we can fetch from a local
webserver. However, this results in a CORS error.

To address this, we can have the Rust side perform the fetch instead of
the webview.

This change replaces the webview fetch with the Tauri fetch API to
enable fetching from a local webserver.